pyproject.toml
[tool.poetry]
name = "foody-backend"
version = "0.1.0"
description = ""
authors = ["Soufiane BEN ZAOUI <soubenz94@gmail.com>"]
readme = "README.md"
packages = [{include = "benz_scraper"}]

[tool.poetry.dependencies]
python = "^3.10"
parsel = "^1.8.1"
pydantic = "1.10.13"
dateparser = "^1.2.0"
replicate = "^0.22.0"
playwright = "^1.41.0"
loguru = "^0.7.2"
playwright-stealth = "^1.0.6"
alembic = "^1.13.1"
sqlalchemy = "^2.0.25"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
aerich = "^0.7.2"
asyncpg = "0.24.0"
python-dotenv = "^1.0.1"



[tool.aerich]
tortoise_orm = "benz_scraper.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
